move norm to norm.cpp file
